better structure for sketches
