+++ demo/main.c
@@ -7,41 +7,8 @@
 enum render_mode { IMAGE, GIF, VIDEO, SMALL_VIDEO } mode;
 double width, height;
 
-////////////////////////////////////////
-// SETUP
-////////////////////////////////////////
-
-bl_render_config setup() {
-  // IMAGE, GIF, VIDEO, SMALL_VIDEO
-  mode = GIF;
-
-  // width, height, frames, fps
-  bl_render_config configs[4] = {
-      {800, 800, 0, 0},     // IMAGE
-      {400, 400, 60, 30},   // GIF
-      {1920, 1080, 60, 30}, // VIDEO
-      {640, 360, 60, 30},   // SMALL_VIDEO
-  };
-  return configs[mode];
-}
-
-////////////////////////////////////////
-// RENDER
-////////////////////////////////////////
-
-void render(cairo_t *cr, double percent) {
-  // percent will be 1.0 when rendering an image
-  cairo_clear_rgb(cr, 1, 1, 1);
-  cairo_set_line_width(cr, 0.5);
-
-  for (int i = 0; i < 20; i++) {
-    cairo_line_to(cr, g_random_double_range(0, width * percent),
-                  g_random_double_range(0, height));
-  }
-  cairo_stroke(cr);
-
-  // where all the magic happens!
-}
+bl_render_config setup();
+void render(cairo_t *cr, double percent);
 
 ////////////////////////////////////////
 // MAIN
