+++ makefile
@@ -9,10 +9,15 @@ default: bin/main
 	@echo running main...
 	@./bin/main
 
-bin/main: libs/bitlib.a demo/main.c
+bin/main: libs/bitlib.a demo/main.c sketch.o
 	@echo compiling main...
 	@mkdir -p bin
-	@$(CC) $(WARN) $(CFLAGS) demo/main.c -Iinclude libs/bitlib.a -o bin/main -lm $(CLIBS)
+	@$(CC) $(WARN) $(CFLAGS) demo/main.c -Iinclude libs/bitlib.a sketch.o -o bin/main -lm $(CLIBS)
+
+sketch.o: demo/sketch.c
+	@mkdir -p build
+	@$(CC) $(WARN) $(CFLAGS) -c demo/sketch.c -Iinclude -o sketch.o -lm $(CLIBS)
+
 
 dist: libs/bitlib.a
 	@echo assembling distribution package for version $(VERSION) ...
@@ -40,3 +45,4 @@ clean:
 	@rm -f *.gif
 	@rm -f *.mp4
 	@rm -f v*.zip
+	@rm -f *.o
